% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQRNN.R
\name{BQRNN}
\alias{BQRNN}
\title{BQRNN}
\usage{
BQRNN(Y, X, k, tau = 0.5, prior, stepsize, m, initialvalue, burnin, l)
}
\arguments{
\item{Y}{the response variables}

\item{X}{the predictive variables}

\item{k}{the number of hidden layer nodes}

\item{tau}{quantile}

\item{prior}{hyper-parameters in priors}

\item{stepsize}{the variance of proposal distribution in random walk Metropolis–Hastings algorithm}

\item{m}{the length of  MCMC chains}

\item{initialvalue}{initial values of parameters}

\item{burnin}{burn in of MCMC chains}

\item{l}{step size of MCMC chains}
}
\value{
the estimates of parameters and MCMC diagnostic statistics\code{n}
}
\description{
An implementation of Bayesian Quantile Regression Neural Networks model using R
}
\examples{
\dontrun{
       x<-matrix(stats::runif(600,0,5),200,3)
       beta1<-c(2,4,6)
       beta2<-c(0.1,0.3,0.5)
       e<-stats::rnorm(200,0,1)
       y<-x\%*\%beta1+x\%*\%beta2*e
       k<-4
       p<-3
       prior<-list(rep(0,k+1),10,matrix(0,p+1,k),10,3,0.1)
       stepsize<-0.01^2
       m<-1e5
       initialvalue<-list(rep(0,k+1),matrix(0,p+1,k))
       burnin<-m/2
       l<-10
       a<-BQRNN(y,x,k,0.05,prior,stepsize,m,initialvalue,burnin,l)
       a
}
}
