% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmnGibbsc}
\alias{rmnGibbsc}
\title{A Gibbs sampler for multivariate normal distribution using Rcpp}
\usage{
rmnGibbsc(m, x0, mu, sigma)
}
\arguments{
\item{m}{the number of samples}

\item{x0}{the initial value}

\item{mu}{the mean vector of the distribution}

\item{sigma}{the covariance matrix of the distribution}
}
\value{
a MCMC chain \code{n}
}
\description{
A Gibbs sampler for multivariate normal distribution using Rcpp
}
\examples{
\dontrun{
a<-rmnGibbsc(2, c(1,2,3),c(0,0,0),matrix(c(1,0.5,0.5,0.5,1,0.5,0.5,0.5,1),3,3))
a
}
}
