\name{SA23204191-package}
\alias{SA23204191-package}
\alias{SA23204191}
\docType{package}
\title{
\packageTitle{SA23204191}
}
\description{
Rcpp version of Gibbs sampler for multivariate normal distribution and an implementation of Bayesian Quantile Regression Neural Networks model.
\packageDescription{SA23204191}

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SA23204191}
\packageIndices{SA23204191}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
zixiao Ma\packageAuthor{SA23204191}
Maintainer:zixiao Ma <mazixiao@ustc.mail.edu.cn> \packageMaintainer{SA23204191}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }

